+++ src/main/java/kotlinx/reflect/lite/impl/ConstructorMetadataImpl.kt
@@ -19,6 +19,7 @@ package kotlinx.reflect.lite.impl
 import kotlinx.reflect.lite.ConstructorMetadata
 import kotlinx.reflect.lite.DeclarationMetadata
 import kotlinx.reflect.lite.ParameterMetadata
+import kotlinx.reflect.lite.TypeMetadata
 import org.jetbrains.kotlin.serialization.Flags
 import org.jetbrains.kotlin.serialization.ProtoBuf
 import org.jetbrains.kotlin.serialization.deserialization.NameResolver
@@ -32,6 +33,9 @@ internal class ConstructorMetadataImpl(
 
     override val parameters: List<ParameterMetadata>
         get() = proto.valueParameterList.map { ParameterMetadataImpl(it, nameResolver) }
+
+    override val extensionReceiverType: TypeMetadata?
+        get() = null
 
     override val isPrimary: Boolean
         get() = !Flags.IS_SECONDARY.get(proto.flags)