+++ CHANGELOG.md
@@ -4,8 +4,7 @@
 
 * Support property metadata with ClassMetadata.getProperty
 * Allow to enumerate all declarations in a class with ClassMetadata.{functions,constructors,properties}
-* Add CallableMetadata.name
-* Add CallableMetadata.returnType
+* Add CallableMetadata.{name,returnType,extensionReceiverType}
 * Add ClassMetadata.isData
 * Add ClassMetadata.kind
 * Add ConstructorMetadata.isPrimary